Remove misc.py
