--- rampwf/utils/misc.py
@@ -1,49 +1,0 @@
-# coding: utf-8
-"""
-Miscalleneous methods
-"""
-import os
-
-import cloudpickle as pickle
-
-from .colors import print_warning
-
-
-def delete_line_from_file(f_name, line_to_delete):
-    with open(f_name, "r+") as f:
-        lines = f.readlines()
-        f.seek(0)
-        for line in lines:
-            if line != line_to_delete:
-                f.write(line)
-        f.truncate()
-
-
-def pickle_model(fold_output_path, trained_workflow, model_name='model.pkl'):
-    """Pickle and reload trained workflow.
-
-    If workflow can't be pickled, print warning and return origina' workflow.
-
-    Parameters
-    ----------
-    fold_output_path : str
-        the path into which the model will be pickled
-    trained_workflow : a rampwf.workflow
-        the workflow to be pickled
-    model_name : str (default='model.pkl')
-        the file name of the pickled workflow
-    Returns
-    -------
-    trained_workflow : a rampwf.workflow
-        either the input workflow or the pickled and reloaded workflow
-    """
-    try:
-        model_file = os.path.join(fold_output_path, model_name)
-        with open(model_file, 'wb') as pickle_file:
-            pickle.dump(trained_workflow, pickle_file)
-        with open(model_file, 'r') as pickle_file:
-            trained_workflow = pickle.load(pickle_file)
-    except Exception as e:
-        print_warning("Warning: model can't be pickled.")
-        print_warning(e)
-    return trained_workflow
