+++ rampwf/utils/notebook.py
@@ -7,7 +7,15 @@ from __future__ import print_function
 import os
 import subprocess
 
-from .misc import delete_line_from_file
+
+def delete_line_from_file(f_name, line_to_delete):
+    with open(f_name, "r+") as f:
+        lines = f.readlines()
+        f.seek(0)
+        for line in lines:
+            if line != line_to_delete:
+                f.write(line)
+        f.truncate()
 
 
 def execute_notebook(ramp_kit_dir='.'):
